+++ packages/db-chuni/src/schema/types.ts
@@ -1,6 +1,6 @@
 import { pgEnum } from "drizzle-orm/pg-core";
 
-import { ranks } from "@repo/utils-chuni";
+import { ranks } from "@repo/utils/chuni";
 
 export const categoryValues = [
   "POPS & ANIME",