+++ packages/utils/src/chuni/index.ts
@@ -1,4 +1,4 @@
-export * from "./decimal";
+export * from "./level";
 export * from "./rank";
 export * from "./rarity";
 export * from "./rating";